Fix issue 379 - Update ionic seed project
Both samples in this repo are due for an update. This includes updated
dependencies for both, as well as remove the contents of the platform
and plugin directories, since these would be generated in the local
copy of the project during the first build or when adding the ionic
platform (iOS or android).