Updates package references versions
This updates the package.son file on both examples to make it
compatible with the current Ionic and Nodejs. Also, a step to add the
iOS platform to the Ionic project was missing on both examples
README.md instructions, so it was added. Also the examples versions
were upgraded to 1.0.1 .

Fixes https://github.com/auth0/auth0-ionic/issues/24